Create website draft for puzzle (#2)
This PR creates the website and creates the buttons.

With this PR, there is also the skeleton of the TypeScript code which allows the user to play around with the puzzle (it has no additional features). The TypeScript has NOT been tested nor has it been compiled to JavaScript yet.